build: Add build scripts
This adds a build script that:

- builds the main Node.js library using `tsc`
- builds the client-side library using Vite
- compiles the Sass

The client-side library is made available under a separate export using the `exports` key in `package.json`. This is a relatively-new feature compatible with Webpack 5 and other bundlers, and with Node.js 16+. It is also supported by TypeScript when `compilerOptions.moduleResolution` is set to `node16` (however, the client-side library doesn't includes types as GOV.UK Frontend doesn't).