feat(Tag): Add support for colour variants
This adds support for the additional colour variants that the GOV.UK Tag component has.

This was tricky as we don't want to apply our override when a colour variant is specified. I added a new Nunjucks custom filter that can be used to add a CSS class depending on what other classes are set, and used this to only apply our override when a custom colour isn't being applied.