feat(footer): Add component
This adds the footer component. There are no changes from the GOV.UK version.

This required having the crest (Royal Coat of Arms) image available. This is selectively copied from the GOV.UK Frontend as a build step. (Other static assets, such as the GDS Transport font and the GOV.UK crown aren't copied, as we aren't using them.)