feat(header): Add component
This adds the header component, based on the GOV.UK equivalent. The original macro in the GOV.UK Frontend is only intended for use on GOV.UK, and isn't that customisable, so our own version has been added here.

I added some specific visual regression tests for hover states, and also with JavaScript disabled.

Two new Playwright tags – @desktop and @mobile – have been set up, to restrict tests to desktop or mobile. (test.skip() was an option, but the tags seemed like a cleaner way to do it.)