feat(fieldset): Add component
This adds the fieldset component. There were no changes from the GOV.UK component.

Note the wrapper macro needed some special handling to ensure it works when a [call block](https://mozilla.github.io/nunjucks/templating.html#call) is used (as well as when the `html` property is used). This special handling needs to be ported to a few other components as well.