ci(release): Release non-alpha versions only
As the library has now gone through internal testing, this makes the release workflow release normal, non-alpha versions. This will eliminate any related friction when testing with users.

The ability to release an alpha version has been removed completely, as it doesn't fit in with the automated release process.

As before, I've given this a quick test in https://github.com/jpveooys/github-actions-testing-v2.