deps-dev: Ensure consistent Playwright version is used
When `@playwright/test` is updated by Dependabot, we tend to have problems as `@axe-core/playwright` has a dependency on `playwright` and the versions get out of sync.

This adds an override to make sure `@axe-core/playwright` uses the same version of Playwright as the project to make updates less painful.