test(e2e): Add tablet project
This replaces the iPhone 12-based project in the Playwright configuration with an iPad (gen 6)-based project.

This allows a tablet screen size to be tested.

`@tablet`, `@mobile-and-tablet`, `@tablet-and-desktop` test tags were added as part of this.

The functionality of the tags can be checked by running a test like:

```
  test.only('@tablet', async ({ page }) => {
    await expect(true).toBeFalsy()
  })
```

and checking which projects it runs for.