build: Include React examples in built package
This updates the `files` and `exports` package.json properties so that the React examples are included in the built package, and can be resolved or imported from the https://github.com/defencedigital/moduk-frontend-guidance-site/.

It also makes the React examples use external imports so that they can serve as standalone examples and be reused for the code snippets. This required configuring aliases in Webpack and TypeScript.