feat(header): Update active navigation item colour
This updates the colour of the active navigation item link to match the current designs.

The [colour threshold](https://playwright.dev/docs/api/class-locatorassertions#locator-assertions-to-have-screenshot-1-option-threshold) in the visual regression tests was also reduced from the default of `0.2` to `0.05`, because the change to the active navigation item colour wasn't being picked up with `0.2`. This has caused some other insignificant changes to be picked up.

`svgo` was also disabled for `babel-plugin-inline-react-svg` as it was causing small rendering differences between React and Nunjucks. There was some refactoring of the Babel configuration to make more of it shared between Eleventy and the production build (making use of [config merging](https://babeljs.io/docs/configuration#how-babel-merges-config-items)).

`footer-max-line-length` was also disabled in the commitlint config as it was triggering incorrectly.