feat(react): Include React components in built package
This makes the built React components (accordion, back link and header) importable from `@moduk/frontend/react` in the published package.

This changes how the library is packaged by:

- packaging from the `dist` directory
- running the `build:react` script as part of the `build` script
- adding `react` entries to the `exports` package.json property

Packaging from the `dist` directory is more complex, but makes it possible to make importing `@moduk/frontend/react` work regardless of environment or framework. If the `exports` field is supported by the importing environment, it will select an ESM or CommonJS build as appropriate. If the `exports` field is not supported, the CommonJS build will be imported as it is located at that path.

For backwards compatibility, the previous `dist` directory is also included in the built package. This can be removed in the next major version.

This also moves more things under `src/react/internal` and adds marks all React components as experimental using a comment.