feat(phase-banner): Add React implementation
This adds a React implementation of the phase banner component.

Because the Nunjucks version allows all attributes of the tag component to be changed, an instance of `PhaseTag` is passed as a prop to `PhaseBanner` in the React version. This allows customisation of the tag component in a relatively clean manner.

The default tag Nunjucks examples was also cleaned up, to remove properties set only for testing.