feat(input): Add React implementation
This adds a React implementation of the text input component, including the supplementary components (error message, hint, label etc).

Similar to some other components, instances of those supplementary components are passed as props. This allows customisation of them in a clear way (and still using JSX), while still giving the root component some control and awareness of them.

The `id` for the `<input>` and related elements are automatically generated using `useId()` if an `id` is not explicitly specified. The `id` is passed to child components using a context so that it doesn't need to be repeated or passed around in awkward ways.