feat(table): Add React implementation
This adds a React implementation of the table component. The table must be composed using children, rather than the table data being passed through props. This does mean a lot of components are exposed, but these simply reflect the underlying HTML elements.

`firstCellIsHeader` is not exposed as a prop, instead `TableRowHeader` is provided to use for the first cell in a row when desired.