feat(notification-banner): Add React implementation
This adds a React implementation of the notification banner component.

The JavaScript for the component is relatively simple, so it's been reimplemented.

There is a subtlety with `children` as the original component wraps text (and not HTML) in a `<p className='govuk-notification-banner__heading'>`. This is done here when `children` doesn't contain React elements. Arrays of strings (e.g. `text{" "}` is accounted for in this.

The title of the banner is not exposed as a component, as the original Nunjucks component does not allow much customisation of it (it's not an object there like other components).