test(e2e): replace `tablet-safari` tests with `tablet-chrome`
An update to `@playwright/test` caused visual regression tests in Safari to occasionally fail.

Therefore the tablet tests, previously using the iPad (gen 6) device, were changed to use the Galaxy Tab S9, which uses a chromium browser by default.

Ideally we should be testing every component in every browser on desktop, tablet, and mobile. This should be done in the future when there is time to debug the Safari issue.